Replaced size_t by uint64 in encodeuint64.cpp

+++ utilities/encodeuint64.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 int main(int argc, char* argv[])
 {
-	size_t g;
+	uint64_t g;
 	const auto a = reinterpret_cast<char*>(&g);
 	const auto s = sizeof(g);
 	for (string line; getline(cin, line);)
